# Added benchmark + improved Calc performance.