Hello,

This PR improves the performance of the `Calc` function, roughly halving the time it takes and reduces the allocations from 7 to 3, see bellow for the benchstat output. The main speedup has been from creating a buffer that the community ID parameters are written into, before passing them all to the sha1's `Write` call. 

I did take a shortcut and use the [samber/lo](https://github.com/samber/lo) package (Specifically, the [TernaryF](https://github.com/samber/lo#ternaryf) function). If you would prefer that the imports of this package stay minimal, then I am happy to remove it! 😄 

```
goos: darwin
goarch: arm64
pkg: github.com/satta/gommunityid
             │ master.txt  │           this_branch.txt           │
             │   sec/op    │   sec/op     vs base                │
Calc/IPv4-10   309.9n ± 0%   154.7n ± 0%  -50.08% (p=0.000 n=15)
Calc/IPv6-10   436.4n ± 0%   142.8n ± 0%  -67.28% (p=0.000 n=15)
geomean        367.8n        148.6n       -59.58%

             │ master.txt │           this_branch.txt           │
             │    B/op    │    B/op     vs base                 │
Calc/IPv4-10   152.0 ± 0%   152.0 ± 0%       ~ (p=1.000 n=15) ¹
Calc/IPv6-10   174.0 ± 0%   184.0 ± 0%  +5.75% (p=0.000 n=15)
geomean        162.6        167.2       +2.83%
¹ all samples are equal

             │ master.txt │          this_branch.txt           │
             │ allocs/op  │ allocs/op   vs base                │
Calc/IPv4-10   7.000 ± 0%   3.000 ± 0%  -57.14% (p=0.000 n=15)
Calc/IPv6-10   7.000 ± 0%   3.000 ± 0%  -57.14% (p=0.000 n=15)
geomean        7.000        3.000       -57.14%
```