@satta 2023-09-05
All fine with me, thanks for this contribution and for the considerable speedup! As far as dependencies go, `lo` is acceptable.